# students/k3340/laboratory_works/Bogdanov_Maxim/Lr3/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_delivery_http_handler.CreateTeacherRequest:
    properties:
      classroom_id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
    type: object
...
  internal_delivery_http_handler.TeacherResponse:
    properties:
      classroom_id:
        type: integer
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_delivery_http_handler.UpdateTeacherRequest:
    properties:
      classroom_id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@school-service.com
    name: API Support
  description: API для управления школьной системой
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: School Management Service API
  version: "1.0"
paths:
...
  /teachers:
    get:
      consumes:
      - application/json
      description: Возвращает список всех учителей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http_handler.TeacherResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список учителей
      tags:
      - teachers
    post:
      consumes:
      - application/json
      description: Создает нового учителя в системе
      parameters:
      - description: Данные учителя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.CreateTeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.TeacherResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать учителя
      tags:
      - teachers
  /teachers/{id}:
    delete:
      consumes:
      - application/json
      description: Выполняет soft delete учителя
      parameters:
      - description: ID учителя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить учителя
      tags:
      - teachers
    get:
      consumes:
      - application/json
      description: Возвращает информацию об учителе по его ID
      parameters:
      - description: ID учителя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.TeacherResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить учителя по ID
      tags:
      - teachers
    put:
      consumes:
      - application/json
      description: Обновляет информацию об учителе
      parameters:
      - description: ID учителя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные учителя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.UpdateTeacherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http_handler.TeacherResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить учителя
      tags:
      - teachers
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"